html/semantics/scripting-1/the-script-element/scriptexecute-events/013.html

<!DOCTYPE html>
<html>
  <title>document.write inline script in beforescriptexecute</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <div id="log"></div>
  <script>
    targets = [];
    seq = [];
    var t = async_test();
    var a = 0;
    document.addEventListener("beforescriptexecute",
      function x(e) {
       targets.push(e.target);
       if (a <= 1000) {
         document.write("<script>seq.push(" + (a++) + ")<\/script>");
       } else {
          document.removeEventListener("beforescriptexecute", x);
       }
    }, false);
 </script>
 <script>var a = 1</script>
 <script>
 t.step(function(){
   var expected = [];
   for (var i=1000; i>=0; i--) {expected.push(i);}
   assert_equals(seq, expected);
 });
 t.done();
 </script>
